information_leaks_map/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
flake8>=6.0,<7.0
python-telegram-bot>=20.0,<21.0
bokeh>=3.0,<4.0
reportlab>=4.0,<5.0
pandas>=2.0,<3.0
python-dateutil>=2.8,<3.0
aiohttp>=3.8,<4.0
